[DPE-5762] Add CI workflows (#2)
# Summary
In order to build, test and publish the etcd operator automatically, we will run Github action workflows when creating or pushing to PR's or merging PR's to the default branch. These workflows will be based on data-platform-workflows.

# Note
We are currently still using:
- Ubuntu 22.04 base (due to Add support for building on 24.04 data-platform-workflows#246, data-platform-workflows don't support 24.04 yet)
- Juju 3.5.4 as default for integration testing (3.6 has not yet been released to stable, but is already added to scheduled test runs)

Furthermore the release job for publishing to Charmhub is not active yet because credentials still need to be added there.