[DPE-5764] handle config as dict (#4)
This PR improves the handling of config properties. Instead of using a
list of strings, the config is loaded from a yaml-file and handled as a
dict. This allows for controlling all config properties as part of the
code while making it easier to add or update config properties. It also
provides a cleaner approach.

Key changes:
- move config properties to a separate file
- load config properties from yaml to dict instead of from string to
list of strings
- `config_properties` now returns a string instead of `list[str]` which
makes it easier to write to a file