[DPE-6075] Make continuous writes independent of specific units (#19)
The continuous writes function in the integration test suite needs to be
able to run independently. Currently it is executed via the `etcdctl`
command being run in one of the deployed juju units. This is problematic
as we might want to remove this specific unit as part of the scaling/HA
tests.

This PR adjusts integration test setup and installs the deb-package
`etcd-client` on the host machine. The continuous writes function is
then run with this client's `etcdctl` and thus not depending on
execution in specific units anymore.

The function is still kept very simple:
- It needs to be started with `start_continuous_writes(endpoints=...)`
and stopped with `stop_continuous_writes()`.
- It can be started with multiple endpoints though and persists through
changes of Raft or Juju leadership.
- Use the `assert_continuous_writes_increasing` and
`assert_continuous_writes_consistent` helpers to ensure that the writes
continued throughout any cluster/HA related operation.

**Note:**
There is no currently maintained Python client lib for etcd (see
[issue](https://github.com/etcd-io/etcd/issues/13285)), that's why we
keep using `etcdctl` to connect to etcd on each write. Testing with
several of the available, unsupported clients failed because they either
don't support authentication or cannot be installed to due
incompatibilities with protobuf.