Clean tox etcd install (#40)
This pull request includes changes to the `tox.ini` file to improve the
installation process of `etcdctl` by using a temporary directory for
downloads and extraction. This change enhances the script's reliability
and cleanliness.

Improvements to the installation process:

*
[`tox.ini`](diffhunk://#diff-ef2cef9f88b4fe09ca3082140e67f5ad34fb65fb6e228f119d3812261ae51449L81-R84):
Changed the installation commands for `etcdctl` to use a temporary
directory for downloading and extracting the archive, ensuring a cleaner
and more reliable installation process.