connect to all cluster endpoints when broadcasting peer URL (#41)
When broadcasting an updated peer URL during the peer TLS transition, it
could happen that connection errors appear when connecting to the local
instance with TLS (because client cert files have been removed by the
separate client TLS disabling hook). This has been observed
[here](https://github.com/canonical/charmed-etcd-operator/actions/runs/13258167474/job/37008789433?pr=36).

To work around that, we could connect to all endpoints instead of just
the local one.