[DPE-6342] trigger leader transfer on restart (#36)
When restarting a member of an etcd cluster, the cluster could
experience a short unavailability if the member is the current raft
leader of the cluster. Until a new leader was elected (or rather: for
the notice period of the other cluster members), no write operation can
be performed. To avoid this, we should conduct a leader-transfer on the
etcd cluster before initiating a restart on the current raft leader.
This PR provides implementation and test coverage for this.

The following changes are introduced:

**ClusterManager:**
- add a new method `move_leader_if_required()` that checks if the member
is currently the leader and, if so, transfers leadership to another
cluster member
- run the `move_leader_if_required()` method as part of the
`restart_member()`

**Restart callback functions in charm.py:**
- when enabling or disabling peer TLS: execute
`move_leader_if_required()` before broadcasting the member's updated
peer URL (fix for
https://github.com/canonical/charmed-etcd-operator/issues/37)
- all other cases: execute `move_leader_if_required()` directly before
restarting the workload (as part of `cluster_manager.restart_member()`)

**Integration tests:**
- add a new test `test_ha_on_rolling_restart.py` to ensure continuous
writes on restarts when enabling or disabling peer TLS
- add a logfile output to continuous writes for easier debugging missed
writes
- move away from the `wait_for_cluster_formation()` helper introduced in
https://github.com/canonical/charmed-etcd-operator/pull/28 (no longer
necessary because we can use `wait_until`)
- adjust certificate validity from 1m to 3m in
`test_certificate_expiration` because in some cases the rolling restart
did not complete before the certificate expired (test results
flickering)

---------

Co-authored-by: Smail Kourta <smail.kourta@canonical.com>