[DPE-6539] Enable storage reuse (#38)
Re-using previously used storage volumes in `charmed-etcd` is possible
under the following conditions:
- An existing database in
`/var/snap/charmed-etcd/common/var/lib/etcd/member` can be used with a
new `charmed-etcd` application or an existing application that is scaled
to zero units.
- There is no way of using an existing database with previous membership
information, but new IP address. Therefore, the cluster membership data
has to be purged/renewed on startup of the first cluster member. This is
because membership information is also stored in the database itself;
when loading the existing database, membership information is also
loaded/overwritten.
- It is not possible to use an existing database for unit `n+1` and only
catch up on the diff from the leader. Additional units/cluster members
can only be added by transferring the complete database from the leader.
- When deploying a new `charmed-etcd` application with an existing
database/existing storage, the previously used root password has to be
supplied by the user, before deploying etcd. This is required because
the password is stored in the database itself, and if not provided, the
`charmed-etcd` operator can not perform its operations on the database.
The password has to be provided as a secret and configured to the
operator via the `system_users` config option.

This PR includes the necessary changes to enable storage reuse in etcd.

**EtcdEvents**
- add event handler for `storage_attached`: fix the permissions of the
data directory to be able to re-install the snap on existing storage
- `on_start`: if it is a new cluster with an existing database, start
the first member and update its membership information (peer-url)
- `on_start`: if the cluster is already existing and there is an
existing database in the storage volume, remove the database and start
the new unit/member after full data transfer from the leader
- `on_leader_elected`: only generate a new root password if there is no
user-provided root password configured

**ConfigManager:**
- introduce new config parameter `force-new-cluster` for re-using an
existing database/storage

**Workload:**
- add method `exists()` to check if there already is a database on the
storage volume
- add method `remove_directory()` to purge an existing database if
required
- add method `exec()` to perform directory permission/ownership changes
on the filesystem

**Integration tests:**
new test suite `test_storage_reuse.py` with the following tests:
- `test_attach_storage_after_scale_down`
- `test_attach_storage_after_scale_to_zero`
- `test_attach_storage_after_removing_application`