[DPE-6568] Raise error if enabling auth fails (#48)
Currently the etcd operator does not raise an error if enabling
authentication on cluster startup fails. This could happen for example
if the `etcdctl` command times out because of high load on the machine.

To make sure we do not start an unprotected cluster, the `on_start` hook
should fail if authentication can not be enabled.

The changes in this PR include:

**EtcdEvents:**
- fail the `on_start` hook for the first unit in the cluster, if the
admin user can not be created or authentication can not be enabled in
the cluster
- fail the `on_start` hook for subsequent units if authentication is not
enabled

**Unit tests:**
- add new test cases for unhappy path (when enabling authentication
fails) on startup