[DPE-6790] - TLS Status Handling (#53)
This pull request includes several changes focused on improving the
handling and computation of TLS statuses within the `etcd` charm. The
changes involve updating the methods for setting and computing TLS
statuses, as well as modifying the status levels and their logging
levels.

Improvements to TLS status handling:

*
[`src/charm.py`](diffhunk://#diff-b9ed39bbc9c0387bd3e07da31d13373745534a1cd723d3e292c73496b12e307cR224-R225):
Added logic to compute and add TLS component statuses in the
`_on_collect_status` method.
*
[`src/events/etcd.py`](diffhunk://#diff-4d4b45129d72bf79a08e70009a59f4ebe0a016b4727117155249c864ae4a1b49L106):
Removed the direct setting of TLS status and replaced it with deferring
events and adding component statuses through the `tls_manager`.
[[1]](diffhunk://#diff-4d4b45129d72bf79a08e70009a59f4ebe0a016b4727117155249c864ae4a1b49L106)
[[2]](diffhunk://#diff-4d4b45129d72bf79a08e70009a59f4ebe0a016b4727117155249c864ae4a1b49L363-R362)
*
[`src/events/tls.py`](diffhunk://#diff-3dfb4a104b407f7d4b2e07795a63745a09cde228a8dbe86602e7b72ea600fbe4L99-R105):
Updated the initialization and event handling methods to use the
`tls_manager` for adding component statuses instead of directly setting
the status.
[[1]](diffhunk://#diff-3dfb4a104b407f7d4b2e07795a63745a09cde228a8dbe86602e7b72ea600fbe4L99-R105)
[[2]](diffhunk://#diff-3dfb4a104b407f7d4b2e07795a63745a09cde228a8dbe86602e7b72ea600fbe4L159-L162)

Status level modifications:

*
[`src/literals.py`](diffhunk://#diff-959ba637043ab4ac88c6f9f9109022f78a7df1b9a2ab3a0407bff4f124c2db64L86-R90):
Changed the status level for `TLS_NOT_READY` to `MaintenanceStatus` with
a "DEBUG" log level and added new statuses for peer and client CA
rotations.

Enhancements to `tls_manager`:

*
[`src/managers/tls.py`](diffhunk://#diff-a2c036f44ae9bbab595aad84b23bc3106a342c5ab12f7b0cf3d26268b51cdc83L17-R17):
Introduced a list to track component statuses, and added methods to
compute component statuses and add new statuses to the list.
[[1]](diffhunk://#diff-a2c036f44ae9bbab595aad84b23bc3106a342c5ab12f7b0cf3d26268b51cdc83L17-R17)
[[2]](diffhunk://#diff-a2c036f44ae9bbab595aad84b23bc3106a342c5ab12f7b0cf3d26268b51cdc83R29)
[[3]](diffhunk://#diff-a2c036f44ae9bbab595aad84b23bc3106a342c5ab12f7b0cf3d26268b51cdc83R205-R236)

---------

Co-authored-by: reneradoi <rene.radoi@canonical.com>