[DPE-6706] - Add Client Relation docs (#62)
### Documentation Enhancements:

*
[`docs/how-to/client-relations.md`](diffhunk://#diff-9d69e9be4a9d29b62682ea8b654a5798f6281012b632109697c32dbccf702220R1-R160):
Added a new guide on integrating etcd with an application, including
instructions for using the `etcd_client` interface and the
`data-integrator` charm. This guide covers configuration steps, code
implementation, and commands for establishing and removing relations.
*
[`docs/how-to/index.md`](diffhunk://#diff-a62277b14aab6742697a72ebe63db2be1ea5ba095495f9c7a425d9fc7b6e8f18R12):
Updated the index to include a link to the new client relations
integration guide.