Integrate sphinx starter pack (#61)
This PR adds Canonical's Sphinx tooling to our documentation so that it
can be rendered as a Read The Docs project, and edits the existing docs
to remove build errors and polish the structure.

### Integration with the [Canonical Sphinx Starter
Pack](https://github.com/canonical/sphinx-docs-starter-pack)

We add several files that enable our docs to be built with Sphinx,
checked in our CI, and published on Read The Docs.

In the root directory:
* `.readthedocs.yaml`: Build configuration for the Read The Docs project

Of note in the `/docs` folder:
* `/.sphinx` folder: Build checks and tooling for links, spelling,
inclusive language, and other utilities.
* `conf.py`: Sphinx configuration options, e.g. project name, associated
links, features we want to enable/disable, HTML/CSS, MyST settings,
extensions
* `Makefile`: Makefile for Sphinx's compiler

In the `.github/workflows` folder:
* `automatic-doc-checks.yml`: Reusable workflows from
[canonical/documentation-workflows](https://github.com/canonical/documentation-workflows)
* `markdown-style-checks.yml`: Markdown linter

### Edits to content and structure

The existing files were restructured into Diataxis folders (for now,
just `tutorial` and `how-to`), along with new index pages for each
folder. The tutorial pages were united into a single page.

The content of several pages was polished (i.e. headings, code blocks, and links/refs). 
Several titles and filenames were lightly edited for consistency.

---------

Co-authored-by: Vladimir Izmalkov <48120135+izmalk@users.noreply.github.com>
Co-authored-by: Smail KOURTA <smail.kourta@canonical.com>