[DPE-6780] Implement s3 restore (#55)
This PR adds functionality to restore backups in etcd. 

Backups can be restored to any cluster of `charmed-etcd`, no matter if
the cluster is running or not or how many units are deployed.

The workflow for performing a database restore is the following:
- connect etcd to s3-storage: `juju integrate s3-integrator etcd`
- provide and configure the secret with `root` credentials
- list available backups: `juju run etcd/leader list-backups`
- restore the desired backup, for example: `juju run etcd/leader restore
backup-id=2025-03-28T12:20:43Z`

The etcd operator will then perform these actions on each unit:
- download the backup from s3 storage
- stop and disable etcd 
- restore the backup with updated cluster membership configuration
- enable and start etcd
- delete the downloaded backup file

Details to the orchestration of this workflow are documented on this
wiki page:
https://github.com/canonical/charmed-etcd-operator/wiki/Backup-Restore:-Workflow

The main changes in the code are:

**EtcdClient:**
- re-ordered methods for better readability: `member_list()`,
`is_healthy()`, `broadcast_peer_url()` and `create_database_snapshot()`
moved _before_ the private method `_run_etcdctl()`
- add `_run_etcdutl()` for commands requiring to use `etcdutl` instead
of `etcdctl`
- add `restore_database_snapshot()`

**BackupEvents:**
- add handler for `restore` action
- add handler for `peer_relation_changed` hook; this handler includes
the main logic for orchestrating the restore workflow across all units

**BackupManager:**
- add method `download_backup_file()`
- add methods `stop_database_workload()` and `start_database_workload()`
that not only stop/start etcd but also disable/enable the
systemd-service
- add method `restore_backup()` to perform the actual restore operation
- add method `proceed_restore_workflow_if_possible()` that proceeds the
restore-workflow after all units have completed the current step
- add method `compute_component_status()` for backup-related statuses

**Other event handlers:**
- defer `config_changed` events if restore is in progress
- defer `peer_relation_joined` events to avoid upscaling operation to
the cluster while a restore is in progress
- ignore `peer_relation_changed` if restore is in progress (this event
is then only handled in `BackupEvents`)
- ignore `update_status` if restore is in progress
- `on_storage_detaching`: do not remove members from the etcd cluster if
restore is in progress (because the cluster member configuration will be
updated after the restore anyway)
- defer `certificate_available` and `relation_broken` events for TLS if
restore is in progress

**Unit State:**
- add property `restore_step` for the step of the restore workflow a
unit is currently executing

**Cluster State:**
- add property `restore_instruction` for next step of the restore
workflow that all units have to execute
- add property `restore_id` to store the parameter `backup-id` that is
requested to be restored
- add property `is_restore_in_progress` that can be used to block
operations/events while a backup is restored
- add property `can_restore_workflow_proceed` to check if all units have
completed the current step of the restore workflow

**EtcdWorkload:**
- imports the lib `charms.operator_libs_linux.v1.systemd` for enabling
and disabling the systemd-service for etcd