[DPE-6781] Add azure support (#64)
This PR adds support for Azure object storage as backend for storing
backup files. It is based on top of the
[restore-PR](https://github.com/canonical/charmed-etcd-operator/pull/55)
and merges into the feature-branch for backup/restore implementation.

The implementation is similar to the existing S3-implementation. The
main points of interest are:
- credentials and configuration is shared with etcd via the
`azure-storage-integrator`, implementing the `azure` relation interface
- the `azure_storage` lib is temporarily added to this repo via
copy&paste, as it has not been published in Charmhub yet
- the implementation for azure uses the functionalities provided by
`azure.storage.blob` Python package
- if both the `s3` and `azure` relation is configured, the
relation-events will be handled without raising errors, but the charm
will be in blocked status
- any backup-related user action will fail if both relations are
configured