[DPE-6782] add S3 integration tests (#58)
This PR adds integration tests for creating and restoring a backup with
S3-compatible storage.

The storage is deployed as a microceph instance with rados-gateway
enabled. Please note: The setup of microceph was added as a test fixture
because the solution provided by data-platform-workflows is
[deprecated](https://github.com/canonical/data-platform-workflows/blob/main/python/pytest_plugins/microceph/README.md).

The PR is based on the branch for `s3-restore` and is supposed to be
merged into the feature branch `backup-and-restore-implementation`
**after** the `s3-restore` PR was merged into it.