[DPE-7104] Bugfix: CA rotation fails if client certificates already expired (#81)
fixes https://github.com/canonical/charmed-etcd-operator/issues/70

**Root cause**
If a TLS CA rotation is happening when the TLS client certificates have
expired, the operator is no longer able to communicate with the etcd
cluster when restarting. Health checks fail and cause irrecoverable
errors. This also impacts the `peer_relation_changed` events happening
as part of the CA-rotation workflow.

**Fix**
The methods `_restart_ca_rotation()` and `_restart_clean_cas()` should
ignore Health check errors when restarting etcd. This is safe and
self-healing, as the following steps of the CA-rotation workflow will
also check the health status of the cluster when the operator has
updated the client certificates and can communicate with the cluster
again.

Furthermore, when processing the `peer_relation_changed` event, cluster
communication errors should also be ignored when updating client
relation data (esp. the etcd version).

**Other changes**
- Add a new integration test for CA-rotation triggered by expiring
certificates.
- Add a new method `check_certificate_validity()` to the `TLSManager` to
check whether any of the installed certs/CA-certs expires within the
next 24 hours. This check is executed on `update_status` and sets the
unit and app status to `Blocked - "TLS certificates expiring soon..."`