[DPE-6997] Verify backup before restore (#85)
As part of the restore procedure, the data directories of all units will
be purged. In a situation where the restore fails (e.g. wrong admin
password configured), the cluster would ultimately be lost and all the
data could not be recovered anymore. To avoid this, we will run a
verification of the restore on the leader before actually executing the
restore on all units.

The restore-verification will:
- purge the unit's data directory
- restore the backup file only on the leader unit
- immediately start etcd
- perform a health check on the leader unit

If the health check fails, the restore workflow will NOT proceed, so
that the data on the remaining units will not be purged. Instead, the
leader unit will reset its data directory and the restore-step will be
skipped for all units. The cluster will be recovered as before the
restore was initiated.

If the health check succeeds, the restore procedure can continue on all
units.

**UI/UX:**
- add new status `Blocked: Restore verification failed - etcd cluster
still running, restore cancelled, check debug-log`
- if a new `restore` action is executed by a user, the status will be
reset

**Implementation:**
- add `RestoreStep.VERIFY`
- add new method `_verify_restore()` to the `BackupEvents` class that
bundles the execution flow for the `verify` step
- adjust workflow execution for the `restore` step

**Testing:**
- adjust unit test coverage in `test_restore_workflow_synchronization()`
- add new integration test file `test_restore_verification_failed.py`