[DPE-6998] Recover from majority failure (#95)
If an etcd cluster experiences majority failure, it can not recover on
its own. This PR adds a solution for these kind of error cases. It
allows users to rebuild the existing cluster, but with new cluster
membership configuration, while keeping the existing data.

It includes the following changes:

**UI/UX:**
- add a new `BlockedStatus` if cluster majority failure was detected:
`Cluster failure - majority of cluster members lost. Run action
rebuild-cluster to recover.`
- add a new action `rebuild-cluster` that allows users to recover a
cluster from majority failure be re-initializing the cluster membership
configuration. If the cluster is working fine, the action will fail to
avoid shutting down the cluster unintentionally. This can be overridden
with `force=True`.
- add a new `BlockedStatus` if cluster rebuild was executed by a user:
`Rebuilding with new cluster configuration...`

**Implementation:**
- `EtcdClient`: new method `get_metric()` that allows querying the etcd
metric server for a given metric (used by the `ClusterManager` to detect
cluster failure)
- add check to detected failed clusters on `update_status`
- new handler `_on_rebuild_cluster_action` for `EtcdEvents`
- new properties for unit and cluster state to check the cluster rebuild
progress
- block all potentially concurring event handlers if cluster rebuild is
in progress, such as `config_changed`, `certificate_available` or others
- new method `_rebuild_cluster()` in `EtcdEvents` to orchestrate the
workflow for rebuilding a cluster based on `peer_relation_changed`
events

The workflow is also documented in this wiki page:
https://github.com/canonical/charmed-etcd-operator/wiki/Rebuild-cluster:-Workflow

**Testing:**
- unit test coverage for the `rebuild-cluster` action
- unit test coverage for the workflow for rebuilding the cluster which
is based on `peer_relation_changed` events
- integration test coverage for rebuilding a failed cluster by running
the action in `test_disaster_recovery.py`

Unrelated change:
- updating `protobuf` dependency to `6.31.1` because of
[CVE-2025-4565](https://github.com/canonical/charmed-etcd-operator/security/dependabot/8)