fix: Fix name of release workflow and track key (#105)
This pull request updates the release workflow configuration to use a
different workflow file and modifies the input parameters for clarity
and consistency.

Changes to the release workflow:

*
[`.github/workflows/release.yaml`](diffhunk://#diff-e426ed45842837026e10e66af23d9c7077e89eacbe6958ce7cb991130ad05adaL44-R46):
Updated the workflow to use `release_charm_edge.yaml` instead of
`release_charm.yaml` and replaced the `channel` parameter with `track`
for better alignment with naming conventions.


[Reference](https://github.com/canonical/data-platform-workflows/releases/tag/v32.0.0)