[DPE-7733] fix TLS switchover (#117)
**Issue**

During the TLS switchover for client connections, it might happen that
the operator tries to connect with TLS to etcd servers without TLS (or
vice versa). This can happen due to a mismatch between:
- the client TLS state of another server in the databag at the beginning
of the current event hook
- the actual TLS state of that server at the time an `etcdctl` command
is executed by the operator code

If that server has restarted and changed client TLS mode during the
event execution, the connection from operator code might fail.

This behavior mostly happens when the operator tries to update the peer
url of it's local etcd server when switching TLS mode. It can be seen in
various CI runs, for example
[here](https://github.com/canonical/charmed-etcd-operator/actions/runs/16406465883/job/46361823833#step:13:1172)
and
[here](https://github.com/canonical/charmed-etcd-operator/actions/runs/16432100325/job/46435494512#step:13:2696)
and
[here](https://github.com/canonical/charmed-etcd-operator/actions/runs/16432100325/job/46435494535#step:13:3137).
To mitigate the risk of failed connections, it would be better to only
execute `etcdctl` commands against the local etcd server when updating
the peer url settings.

**Solution**

This PR changes the behavior of the following methods to only execute
the command against the local etcd server instead of all cluster
endpoints:
- `cluster_manager.member`
- `cluster_manager.broadcast_peer_url()`

This reduces the risk of errors for the operator code in situations
where a) client TLS mode is switched over or b) network disruptions
happen.

Due to the changed behavior, it is required to rearrange the operations
executed after a change of IP address (`config_changed`), because the
peer url can only be updated against the local server after the server
has been restarted and picked up its new IP address (it happened before
the restart until this PR).

**Other changes:**
- add error handling for `certificate_available` event in case it is not
possible to get the assigned certificates; this occasionally fails in
integration tests, as can be seen
[here](https://github.com/canonical/charmed-etcd-operator/actions/runs/16406465883/job/46361823622#step:13:19668)
- while working on it: minor adjustments to integration test for
`ca_rotation`:
- less waiting time because secret expiration for expiring certificates
now already happens before the actual expiration (see: [secret
expiration time in TLS
lib](https://github.com/canonical/charmed-etcd-operator/blob/3.6/edge/lib/charms/tls_certificates_interface/v4/tls_certificates.py#L1202-L1204))
  - running on large runner (for performance reasons)