[DPE-7630] make leader/timeout parameters configurable (#119)
In high-latency network environments, it could happen that the fixed
configuration values we currently use for `heartbeat-interval` and
`election-timeout` cause many leader changes. Therefore we should make
these values configurable to users.

**UI/UX changes:**
- introduce config options `election-timeout` and `heartbeat-interval`
- add a `BlockedStatus` in case the config options are set to invalid
values

**Implementation:**
- add `compute_component_status()` method to the `ConfigManager`
- when running the `config_changed` hook, initiate a rolling restart if
the configuration values are valid and different from what is currently
configured
- add a `tuning_parameters_valid` property to the `ConfigManager`; this
checks the validity of the config values and is used for the
`config_changed` event as well as the status computation
- add a `requires_restart()` method to the `ConfigManager` which loads
the current config file from the unit and checks if one of the tuning
parameters was updated
- adjust the `config_properties()` method of the `ConfigManager` to only
apply the config values if they are considered valid
- add a `load_yaml_file()` method to the `EtcdWorkload` class that can
be used to load the current config file and returns a dictionary with
the configuration parameters

**Architecture decisions:**
- config validation is done by a simple method in the `ConfigManager`
instead of introducing a "structured config class" or config model;
reasons: keep it simple, avoid possible future backwards
incompatibilities for config validation
- config validation is executed (and thereby applied) whenever
configuration is written to the unit; this helps to group restarts for
different kind of changes (e.g. config change because of new ip address
after network disruption + adjusted config values)
- `restart_required` validation happens against the current config file
because there is no other way of querying the currently applied values
for the tuning parameters (even though "live" query would be best e.g.
via metrics, but this is not possible)

**Testing:**
- add unit testing for the `config_changed` event
- add integration testing to `test_ha_on_rolling_restart.py` to ensure
the rolling restart is performed maintaining HA when updating these
config values