fix(DPE-7822): Add additional safeguards in update status and creating external clients (#127)
### Cluster state and authentication handling:

*
[`src/events/etcd.py`](diffhunk://#diff-4d4b45129d72bf79a08e70009a59f4ebe0a016b4727117155249c864ae4a1b49L378-R379):
Updated the `_on_update_status` method to ensure the event handler
checks if the cluster state is unavailable (`not
self.charm.state.cluster.cluster_state`) before proceeding.
*
[`src/events/external_clients.py`](diffhunk://#diff-41d3b0e68fba8bc44a83df27841cc482ba1df8a5cfe8e476397765673f46f446R90-R95):
Added a new condition in the `_on_mtls_cert_updated` method to verify
that cluster authentication is enabled
(`self.charm.state.cluster.auth_enabled`) before proceeding. If
authentication is not enabled, an error is logged, the status is updated
to `CLUSTER_NOT_INITIALIZED`, and the event is deferred.