[DPE-7734] [DPE-7786] Add tuning docs and migration guide (#128)
This PR adds user documentation for newly implemented features and a
migration guide to help users migrate to charmed etcd.
- a `How-to` for Tuning in production
- a `How-to` for migrating to charmed etcd

The existing guide for migration an etcd cluster, that is part of the
backup-and-restore documentation, has been renamed. The two guides have
different scopes, the newly added one is a deployment guide for
administrators moving from e.g. legacy etcd to charmed etcd.

The feature PR's are:
- make leader/timeout parameters configurable
https://github.com/canonical/charmed-etcd-operator/pull/119
- add config option for certificate-extra-sans
https://github.com/canonical/charmed-etcd-operator/pull/126