Revert #136 (#137)
This reverts commit 5096962c4244f6f6de9cd73dfd6c2e8702d7f568.

I'm sorry, seems like the changes to the workflow for refresh v3 can not
be merged separately, as they depend on the config files to be present.
Seems like a chicken-and-egg problem and I will have to find a way to
work around it instead.

Edit: I think I've found a workaround by manually adding a tag with the
required format to our previous `3.6/edge` merge, see
[here](https://github.com/canonical/charmed-etcd-operator/releases/tag/v3.6%2F1.0.0).