[DPE-7889] Fix cluster-status computation (#138)
**Issues**

With #124 the computation of cluster statuses has been reworked. This
leads to lots of errors when deploying charmed etcd, because on every
hook, the etcd-internal metrics server is queried for the
`cluster_failed` metric, even though the etcd server has not been
started yet.

Furthermore, if the cluster is not failed, the operator tries to delete
the status, even though it wasn't set as a running status previously.

Another inconsistency is on enabling or disabling peer TLS, where
`Waiting to join cluster` is displayed as a status message even though
the unit is already a cluster member.

**Fixes**

- The check for `cluster_failed` should only be executed if the etcd
server has already been started.
- This status does not have to be deleted as it is not a running status.
- Consider peer TLS transition for checking cluster membership status.