[DPE-7916] Fix restart after vm reboot (#140)
**Issue:**
If a unit is rebooted, the etcd service is not restarted correctly.

**Solution:**
Always restart the etcd service when processing the `on_start` event, no
matter if it was a controlled shutdown (flag `unit_server.is_started`
gets removed) or not.

The PR also adds integration test coverage for rebooting the Raft leader
and all units of a cluster.