[DPE-7735] Upgrades implementation (#139)
This PR adds minor version upgrades without downtime to charmed etcd.

**What the PR includes:**
- the implementation of all functionalities as described in the
[charm-refresh v3
docs](https://canonical-charm-refresh.readthedocs-hosted.com/latest/add-to-charm/)
- a new event handler `MachinesEtcdRefresh` that implements the etcd
specifics for upgrades
- a new class `UpgradesManager` whose only responsibility is to bridge
the gap between `charm-refresh v3` and `advanced statuses` regarding
status handling
- adjustments to `EtcdWorkload` to be able to install a specific snap
revision (not just the latest) and to be able to handle the
`refresh_versions` file from `charm-refresh` instead of literals for
snap revision
- adjustments to critical event handlers to block or defer while upgrade
is in progress (as discussed in the
[spec](https://docs.google.com/document/d/1o5UsdfA-fX_aOireeC-aCE4AVj8GzMWCBpVVouYFCpI/edit?usp=sharing))
- unit test coverage for the specific parts of implementation (not for
the `charm-refresh` functionality itself)

**What is not part of this PR:**
- basic integration test coverage
- edge case integration test coverage

For proper tests, we need to be able to install / upgrade from a charm
version that supports `charm-refresh v3`. Therefore I will create
separate PR's for integration testing, which will then include newer
snap versions as well.