[DPE-7917] Fix TLS attributes for Vault usage (#144)
**Issue**

In order to use Vault as TLS provider, it is required to omit `/` in TLS
names such as common name or DNS sans. Otherwise Vault will not generate
TLS certificates for issued certificate requests.

**Solution**
Adjust the TLS common name and DNS sans to remove the `/`.

This PR also adds: 
- an integration test for using Vault as TLS provider (based on the
draft from @skourta in #77)
- a disclaimer to the docs with explanations about required Vault
version and configuration