fix([DPE-7831]): - Fix integration test for CA rotation (#133)
This pull request updates the CA and certificate rotation integration
test to use longer validity periods and adjusts the waiting logic
accordingly.

## Context
- The cluster would register a full failure upon rotation CA
certificates
- The certificate secret would expire at 90% of the validity of the
certificate, leaving 10% of its lifetime to acquire a new one and do a
rolling restart
- Empirically, a full rolling restart for CA rotation takes around 1m30s

## Problem
- The certificates were expiring before the units acquired the lock,
resulting in the cluster members losing communications with the other
members, which would cause a cluster failure

## Fix
* Increased the CA validity to 20 minutes and certificate validity to 10
minutes.
* Updated the waiting period after CA renewal to 12m to reflect the new
certificate validity settings, and revised the log message to explain
the timing.
* Changed `apps_full_statuses` to `units_full_statuses` in `wait_until`
calls to wait for the idle_time.
[[1]](diffhunk://#diff-a8362a8580830e574ae0f1e2fbd35adbf5ea51ed2ae3a85e78c5a2eebaa4d0c7L243-R250)
[[2]](diffhunk://#diff-a8362a8580830e574ae0f1e2fbd35adbf5ea51ed2ae3a85e78c5a2eebaa4d0c7R282-R292)