pin juju agent-version to ' 3.6.9' because of issue #20714 (#153)
## Description of issue or feature:
Juju release `3.6.10` introduced a regression: New IP addresses are not
picked up after an IP address change, see
[#20714](https://github.com/juju/juju/issues/20714).

## Solution:
Fix the Juju agent-version used in integration tests to the previous
version `3.6.9`, to allow for proper testing and unblock releases.

Therefore we temporarily disable bootstrapping a Juju lxd controller
with concierge and instead do that ourselves in the spread script.

## How was this change tested?
- [ ] Manually
- [ ] Unit tests
- [X] Integration tests