[DPE-8007] Add config option for certificate domain (#155)
## Description of issue or feature:
When using Vault as TLS provider for etcd, we currently only support it
with the configuration `pki_allow_any_name=true`. In order to allow
using Vault (or another TLS provider) with a more restrictive
configuration, we need to provide a specific domain as part of the
common name and DNS SANs in the certificate requests.

[DPE-8007](https://warthogs.atlassian.net/browse/DPE-8007)

## Solution:
The PR adds configuration options for `client-certificate-domain` and
`peer-certificate-domain`. The values set to these configurations will
be appended to the common name and all DNS SANs of the respective
certificate requests, if valid. If invalid, a blocked status will be
displayed to the user.

The existing code structure can mostly be used for this, the foundation
has already been laid when implementing the `certificate-extra-sans`
configuration. Now, only two new methods
`certificate_domain_config_is_valid()` and `build_common_name()` have
been added to the TLS manager.

**Important note:**
Etcd's communication is strictly IP-based, both for peer-to-peer as well
as client-to-server communication. It is not possible to disable IP SANs
in the certificate requests. For more information see: [breaking change
in
v3.2](https://github.com/etcd-io/etcd/blob/main/CHANGELOG/CHANGELOG-3.2.md#breaking-changes)
and [etcd issue 6336](https://github.com/etcd-io/etcd/issues/6336)

The PR also adds user documentation for the new configuration options
and a note regarding IP SANs to the existing TLS docs.

## How was this change tested?
- [X] Manually
- [X] Unit tests
- [X] Integration tests

Manual testing was performed with both self-signed-certificates for
regression testing as well as Vault for ensuring that we are now able to
use Vault as TLS provider with restricted domain names (also covered
with integration test).

[DPE-8007]:
https://warthogs.atlassian.net/browse/DPE-8007?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ